third_party/blink/web_tests/fast/dom/custom/element-upgrade-expected.txt

CONSOLE WARNING: line 17: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: :unresolved pseudo selector is deprecated and will be removed in M80, around February 2020. Please use :not(:defined) instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests the element upgrade algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS host.firstChild instanceof A is true
PASS element instanceof B is true
PASS callCount is 1
PASS element instanceof C is false
PASS Object.getPrototypeOf(element) is HTMLElement.prototype
PASS host.firstChild instanceof D is false
PASS document.querySelector(":unresolved") is host.firstChild
PASS upgradedOrder is ["e1","e2","e3","e4","e5"]
PASS successfullyParsed is true

TEST COMPLETE
